Initial commit for overhaul of the codebase.
