+++ client/.eslintrc.json
@@ -1,0 +1,81 @@
+{
+    "env": {
+        "browser": true,
+        "es6": true,
+        "jest": true
+    },
+    "globals": {
+        "process": true,
+        "module": true
+    },
+    "plugins": [
+        "react"
+    ],
+    "parser": "babel-eslint",
+    "parserOptions": {
+        "ecmaFeatures": {
+            "experimentalObjectRestSpread": true,
+            "jsx": true
+        },
+        "sourceType": "module"
+    },
+    "extends": "airbnb",
+    "rules": {
+        "max-len": [
+            "error",
+            180,
+            2
+        ],
+        "import/no-extraneous-dependencies": [
+            "error",
+            {
+                "devDependencies": [
+                    "**/__mocks__/*.js",
+                    "**/*.test.js",
+                    "**/*story.js",
+                    "**/__SampleData.js"
+                ]
+            }
+        ],
+        "no-case-declarations": "off",
+        "class-methods-use-this": "off",
+        "no-param-reassign": "off",
+        "react/jsx-filename-extension": "off",
+        "react/prefer-stateless-function": "error",
+        "react/prop-types": "error",
+        "react/forbid-prop-types": "error",
+        "react/require-default-props": "error",
+        "react/no-unused-prop-types": "error",
+        "react/no-multi-comp": "off",
+        "jsx-a11y/no-static-element-interactions": "error",
+        "consistent-return": "error",
+        "no-console": "error",
+        "react/destructuring-assignment": "off",
+        "object-curly-newline": "off",
+        "react/no-access-state-in-setstate": "off",
+        "react/button-has-type": "off",
+        "implicit-arrow-linebreak": "off",
+        "operator-linebreak": "off",
+        "prefer-promise-reject-errors": "off",
+        "jsx-a11y/click-events-have-key-events": "off",
+        "prefer-destructuring": "off",
+        "jsx-a11y/anchor-is-valid": "off",
+        "react/jsx-one-expression-per-line": "off",
+        "function-paren-newline": "off",
+        "react/no-unused-state": "off",
+        "jsx-a11y/label-has-associated-control": "off",
+        "react/default-props-match-prop-types": "off",
+        "jsx-a11y/label-has-for": "off"
+    },
+    "overrides": [
+        {
+            "files": [
+                "**/*.test.js",
+                "**/__mocks__/**.js"
+            ],
+            "rules": {
+                "no-underscore-dangle": "off"
+            }
+        }
+    ]
+}
