+++ client/src/components/HomeComponent.jsx
@@ -1,34 +1,35 @@
-import React, { Component } from "react";
+import React, { PureComponent } from 'react';
 import { Link } from 'react-router-dom';
+import autoBind from 'react-autobind';
 
-class HomeComponent extends Component {
+class HomeComponent extends PureComponent {
   constructor(props) {
     super(props);
     this.state = {
-      count: 5
+      count: 5,
     };
-    // change code below this line
-    this.decrement = this.decrement.bind(this);
-    this.increment = this.increment.bind(this);
-    this.reset = this.reset.bind(this);
-    // change code above this line
+    autoBind(this);
   }
+
   // change code below this line
   increment() {
     this.setState({
-      count: this.state.count + 1
+      count: this.state.count + 1,
     });
   }
+
   decrement() {
     this.setState({
-      count: this.state.count - 1
+      count: this.state.count - 1,
     });
   }
+
   reset() {
     this.setState({
-      count: 0
+      count: 0,
     });
   }
+
   // change code above this line
   render() {
     return (
@@ -36,11 +37,11 @@ class HomeComponent extends Component {
         {/* <button className='inc' onClick={this.increment}>Increment!</button>
           <button className='dec' onClick={this.decrement}>Decrement!</button>
           <button className='reset' onClick={this.reset}>Reset</button> */}
-        <h1>Home Component</h1>
+        <h1>Home PureComponent</h1>
         <Link to="/Counter">Counter</Link>
       </div>
     );
   }
-};
+}
 
 export default HomeComponent;
