--- client/src/components/WorkingWithTheBlockchain.js
@@ -1,21 +1,0 @@
-import React, { Component } from "react";
-import autoBind from 'react-autobind';
-
-class WorkingWithTheBlockchain extends Component {
-    constructor(props) {
-        super(props);
-        this.state = { depositTxHash: null };
-        autoBind(this);
-    }
-
-    render() {
-        return (
-            < div >
-                <div className="loader"></div>
-            </div >
-        )
-
-    }
-}
-
-export default WorkingWithTheBlockchain;
