+++ client/src/utils/getWeb3.js
@@ -1,13 +1,4 @@
-import Web3 from "web3";
-
-// const getWeb3 = async () => {
-//   const provider = await new Web3.providers.HttpProvider(
-//     "http://127.0.0.1:7545"
-//   );
-//   const web3 = new Web3(provider);
-//   console.log("No web3 instance injected, using Local web3.", web3);
-//   return (web3);
-// }
+import Web3 from 'web3';
 
 const getWeb3 = async () => {
   if (window.ethereum) {
@@ -23,56 +14,19 @@ const getWeb3 = async () => {
   } else if (window.web3) {
     //   // Use Mist/MetaMask's provider.
     const web3 = window.web3;
-    console.log("Injected web3 detected.");
+    console.log('Injected web3 detected.');
     return (web3);
   } else {
     const provider = await new Web3.providers.HttpProvider(
-      "http://127.0.0.1:7545"
+      'http://127.0.0.1:7545',
     );
     const web3 = new Web3(provider);
-    console.log("No web3 instance injected, using Local web3.");
+    console.log('No web3 instance injected, using Local web3.');
     return (web3);
   }
-
-
-}
+  return null;
+};
 
 const web3 = getWeb3();
-
-
-
-
-// window.addEventListener("load", async () => {
-// Modern dapp browsers...
-// if (window.ethereum) {
-//   const web3 = new Web3(window.ethereum);
-//   try {
-//     // Request account access if needed
-//     await window.ethereum.enable();
-//     // Acccounts now exposed
-//     resolve(web3);
-//   } catch (error) {
-//     reject(error);
-//   }
-// }
-// Legacy dapp browsers...
-// else if (window.web3) {
-//   // Use Mist/MetaMask's provider.
-//   const web3 = window.web3;
-//   console.log("Injected web3 detected.");
-//   resolve(web3);
-// }
-// // Fallback to localhost; use dev console port by default...
-// else {
-// const provider = new Web3.providers.HttpProvider(
-//   "http://127.0.0.1:7545"
-// );
-// const web3 = new Web3(provider);
-// console.log("No web3 instance injected, using Local web3.");
-// resolve(web3);
-// }
-//   });
-// });
-
 
 export default web3;
