+++ package.json
@@ -14,7 +14,7 @@
     "mocha": "^6.2.0"
   },
   "scripts": {
-    "start": "cd client && npm install && npm run start",
+    "start": "cd client && npm run start",
     "test": "mocha test/AutoETHSavingsAccount"
   },
   "repository": {
